requirements.txt
faker==22.5.1
supabase==2.0.3
python-dotenv==1.0.1
Pillow==10.2.0
gotrue==1.3.1
storage3==0.5.2
postgrest==0.13.2
realtime==1.0.0
